added stats display for word counts and remaining hours; changed GetStatusCountHandler to use DocumentDAO for statistics
--HG--
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/HasTransUnitCount.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/HasTranslationStats.java

+++ flies-war/src/main/java/net/openl10n/flies/webtrans/server/rpc/UpdateTransUnitHandler.java
@@ -106,7 +106,7 @@ public class UpdateTransUnitHandler extends AbstractActionHandler<UpdateTransUni
 
       session.flush();
 
-      long wordCount = hTextFlow.getWordCount();
+      int wordCount = hTextFlow.getWordCount().intValue();
       TransUnitUpdated event = new TransUnitUpdated(new DocumentId(hTextFlow.getDocument().getId()), action.getTransUnitId(), wordCount, prevStatus, action.getContentState());
 
       TranslationWorkspace workspace = translationWorkspaceManager.getOrRegisterWorkspace(action.getWorkspaceId());
